+++ bin/util.py
@@ -37,7 +37,7 @@ def read_or_new_pickle(path: str) -> dict:
     return pickle_dict
 
 
-def older_than(month: int, modified_epoch: int) -> bool:
+def _older_than(month: int, modified_epoch: int) -> bool:
     """
     Determine if a modified epoch date is older than X month
 
@@ -54,47 +54,6 @@ def older_than(month: int, modified_epoch: int) -> bool:
     date = dt.datetime.fromtimestamp(modified)
 
     return date + relativedelta(months=+month) < dt.datetime.today()
-
-
-def validate_directory(directory: str, month: int) -> bool:
-    """
-    Function to check if project (002/003) for that directory
-    exist. e.g. For 210407_A01295_0010_AHWL5GDRXX
-    it will find 002_210407_A01295_0010_AHWL5GDRXX project
-
-    If the 002/003 exist, we check if the proj has been inactive
-    for the last X month. If yes, return True.
-
-    Inputs:
-    :param: directory: trimmed directory pathway
-    :param: month: number of month inactive
-
-    Returns:
-        - `True` if its 002 has not been active for X month
-        - `False` if no 002 project
-        - `False` if 002 project has been active in the past X month
-    """
-
-    result: list = list(
-        dx.find_projects(
-            directory,
-            name_mode="regexp",
-            describe={"fields": {"modified": True}},
-            limit=1,
-        )
-    )
-
-    # if no 002/003 project
-    if not result:
-        return False
-
-    modified_epoch: int = result[0]["describe"]["modified"]
-
-    # check modified date of the 002/003 proj
-    if older_than(month, modified_epoch):
-        return True
-    else:
-        return False
 
 
 def dx_login(today: dt.datetime, token: str, slack: SlackClass) -> None:
@@ -133,33 +92,7 @@ def dx_login(today: dt.datetime, token: str, slack: SlackClass) -> None:
         raise ValueError("Error with DNAnexus login")
 
 
-def remove_project_tag(project_id: str) -> None:
-    """
-    Function to remove tag 'no-archive' for project-id
-
-    When a project has been tagged 'no-archive' but
-    has not been modified for X months
-    the tag will be removed and the project will be
-    notified to be archived
-
-    This only applies to project-id
-    Not directories in Staging52
-
-    Parameters:
-    :param: project-id
-    """
-
-    logger.info(f"REMOVE TAG: {project_id}")
-    try:
-        dx.api.project_remove_tags(
-            project_id,
-            input_params={"tags": ["no-archive"]},
-        )
-    except Exception as e:
-        logger.error(f"REMOVE TAG: {project_id} failed with error {e}")
-
-
-def get_projects_as_dict(project_type: str) -> dict:
+def _get_projects_as_dict(project_type: str) -> dict:
     """
     Function to fetch certain project type and return as
     dict (key: project id, value: describe return from dxpy)
@@ -192,196 +125,24 @@ def get_projects_as_dict(project_type: str) -> dict:
     return result_dict
 
 
-def get_two_and_three_projects_as_single_dict() -> dict:
+def _get_two_and_three_projects_as_single_dict() -> dict:
     """
     Function to get all 002 and 003 projects as a single dict
     """
-    projects_dict_002: dict = get_projects_as_dict("002")
-    projects_dict_003: dict = get_projects_as_dict("003")
+    projects_dict_002: dict = _get_projects_as_dict("002")
+    projects_dict_003: dict = _get_projects_as_dict("003")
 
     return {**projects_dict_002, **projects_dict_003}
 
 
-def get_old_enough_projects(
-    month2: int,
-    month3: int,
-    project_ids_to_exclude: set,
-) -> dict:
+def _add_tag_to_project(tag: str, project_id: str) -> None:
     """
-    Function to get all 002 and 003 projects which are not modified
-    in the last X months.
-
-    Exclude projects: staging 52 as that will be processed separately and
-        exclude projects which had been archived.
+    Add tag to project. Deal with exceptions
 
     Parameters:
-    :param: month2: duration of inactivity in the last x month for 002 (int)
-    :param: month3: duration of inactivity in the last x month for 003 (int)
-    :param: project52: staging-52 project-id
-    :param: project53: staging-53 project-id
-
-    Return:
-    dict of key (proj-id) : value (describe return from dxpy)
-
+    :param: tag: `str` tag to add to project
+    :param: project_id: `str` project id to add tag to
     """
-
-    all_projects = get_two_and_three_projects_as_single_dict()
-
-    projects_that_are_inactive = {
-        k: v
-        for k, v in all_projects.items()
-        if (
-            older_than(
-                month2,
-                v["describe"]["modified"],
-            )  # condition for 002
-            if v["describe"]["name"].startswith("002")
-            else older_than(
-                month3,
-                v["describe"]["modified"],
-            )  # condition for 003
-            and v["describe"]["dataUsage"] != v["describe"]["archivedDataUsage"]
-            and k not in project_ids_to_exclude
-        )
-    }
-
-    # get projects tagged 'archive'
-    projects_tagged_with_archive = list(
-        dx.search.find_projects(
-            name="^00[2,3].*",
-            name_mode="regexp",
-            tags=["archive"],
-            describe={
-                "fields": {
-                    "name": True,
-                    "tags": True,
-                    "dataUsage": True,
-                    "archivedDataUsage": True,
-                    "modified": True,
-                    "createdBy": True,
-                }
-            },
-        )
-    )
-
-    projects_that_are_inactive.update(
-        {proj["id"]: proj for proj in projects_tagged_with_archive}
-    )
-
-    return projects_that_are_inactive
-
-
-def get_all_directories_in_project(project_id: str) -> list:
-    """
-    Function to get all directories in a project-id
-
-    Parameters:
-    :param: project_id: DNAnexus project-id
-
-    Return:
-    list of tuple for ease of processing later on
-    tuple contains:
-        - trimmed directory name (e.g. 210407_A01295_0010_AHWL5GDRXX)
-            for 002 querying later on
-        - original directory path (e.g. /210407_A01295_0010_AHWL5GDRXX/)
-    """
-
-    try:
-        dx_project = dx.DXProject(project_id)
-    except Exception as e:
-        logger.error(e)  # probably wont happen but just in case
-        return []
-
-    # filter out the /processed folder in root of staging-52
-    directories_in_staging_area_52 = [
-        (file.lstrip("/").lstrip("/processed"), file)
-        for file in dx_project.list_folder(only="folders")["folders"]
-        if file != "/processed"  # directories in root of staging-52
-    ] + [
-        (file.lstrip("/").lstrip("/processed"), file)
-        for file in dx_project.list_folder(
-            "/processed",
-            only="folders",
-        )[
-            "folders"
-        ]  # directories in /processed folder
-    ]
-
-    return directories_in_staging_area_52
-
-
-def get_files_in_project_based_on_one_tag(tag: str, project_id: str) -> list:
-    """
-    Function to get files in a project based on a single tag
-
-    Parameters:
-    :param: tag: tag to search for
-    :param: project_id: project-id to search for
-    """
-    if not tag:
-        return []
-
-    results = list(
-        dx.find_data_objects(
-            project=project_id,
-            tags=[tag],
-            describe={
-                "fields": {
-                    "folder": True,
-                },
-            },
-        )
-    )
-
-    return results
-
-
-def get_projects_and_directory_based_on_single_tag(
-    tag: str,
-    project_id: str,
-) -> list:
-    """
-    Function to get all projects and directories tagged with
-    a single tag
-
-    Parameters:
-    :param: tag: tag to search for
-    :param: project_id: project-id to search for
-    """
-
-    if not tag:
-        return []
-
-    results = []
-
-    staging_area_files = [
-        file["describe"]["folder"].lstrip("/")
-        for file in get_files_in_project_based_on_one_tag(
-            tag,
-            project_id,  # stagingarea-52 project id
-        )
-    ]
-
-    # clean the directory path and append to list
-    for directory in staging_area_files:
-        temp = directory.split("/")
-        if "processed" in directory:
-            results.append(f"/{temp[0]}/{temp[1]} in `staging52`")
-        else:
-            results.append(f"{temp[0]} in `staging52`")
-
-    agg_dict = get_two_and_three_projects_as_single_dict()
-
-    results += [
-        proj["describe"]["name"]
-        for proj in agg_dict.values()
-        if tag in proj["describe"]["tags"]
-    ]
-
-    return results
-
-
-def add_tag_to_project(tag: str, project_id: str) -> None:
     try:
         dx.api.project_add_tags(
             project_id,
@@ -400,7 +161,14 @@ def add_tag_to_project(tag: str, project_id: str) -> None:
         logger.error(e)
 
 
-def remove_tags_from_project(tags: list, project_id: str) -> None:
+def _remove_tags_from_project(tags: list, project_id: str) -> None:
+    """
+    Remove tag from project. Deal with exceptions
+
+    Parameters:
+    :param: tags: `list` tags to remove from project
+    :param: project_id: `str` project id to remove tag from
+    """
     try:
         dx.api.project_remove_tags(
             project_id,
@@ -422,7 +190,7 @@ def remove_tags_from_project(tags: list, project_id: str) -> None:
 def tagging_function(debug: bool) -> None:
     """
     Function to check latest archivalStatus of files
-    in a project and add appropriate tag
+    in a project and add appropriate tag to the project
 
     Parameters:
     :param: debug: `bool` whether the script is ran in DEBUG mode
@@ -430,7 +198,11 @@ def tagging_function(debug: bool) -> None:
 
     logger.info("Running tagging function")
 
-    all_projects = get_two_and_three_projects_as_single_dict()
+    if debug:  # if debug, return
+        logger.info("Running in DEBUG mode. Skipping tagging function")
+        return
+
+    all_projects = _get_two_and_three_projects_as_single_dict()
 
     # separate out those with archivedDataUsage == dataUsage
     # which are fully archived so we don't have to query them
@@ -443,17 +215,16 @@ def tagging_function(debug: bool) -> None:
     for project_id, v in archived_projects.items():
         tags = [tag.lower() for tag in v["describe"]["tags"]]
 
-        if not debug:  # if running in production
-            if "partial archived" in tags:
-                remove_tags_from_project(
-                    ["partial archived", "fully archived"],
-                    project_id,
-                )
-                add_tag_to_project(project_id, "fully archived")
-            elif "fully archived" in tags:
-                continue
-            else:
-                add_tag_to_project(project_id, "fully archived")
+        if "partial archived" in tags:
+            _remove_tags_from_project(
+                ["partial archived", "fully archived"],
+                project_id,
+            )
+            _add_tag_to_project(project_id, "fully archived")
+        elif "fully archived" in tags:
+            continue
+        else:
+            _add_tag_to_project(project_id, "fully archived")
 
     # whatever is leftover from above projects, we do the query
     # they can be 'live' or 'partially archived'
@@ -484,45 +255,43 @@ def tagging_function(debug: bool) -> None:
         )
 
         if "archived" in status and "live" in status:
-            if not debug:  # if running in production
-                if "fully archived" in tags:
-                    # if 'fully archived' in tags
-                    # we do a reset and add 'partial'
-                    dx.api.project_remove_tags(
-                        project_id,
-                        input_params={
-                            "tags": ["partial archived", "fully archived"],
-                        },
-                    )
-                    remove_tags_from_project()
-                    add_tag_to_project("partial archived", project_id)
-                elif "partial archived" in tags:
-                    # if 'partially archived' is present
-                    # this project is correctly tagged
-                    continue
-                else:
-                    add_tag_to_project("partial archived", project_id)
+            if "fully archived" in tags:
+                # if 'fully archived' in tags
+                # we do a reset and add 'partial'
+                dx.api.project_remove_tags(
+                    project_id,
+                    input_params={
+                        "tags": ["partial archived", "fully archived"],
+                    },
+                )
+                _remove_tags_from_project()
+                _add_tag_to_project("partial archived", project_id)
+            elif "partial archived" in tags:
+                # if 'partially archived' is present
+                # this project is correctly tagged
+                continue
+            else:
+                _add_tag_to_project("partial archived", project_id)
         elif "live" not in status:
             # everything is archived within the project
-            if not debug:
-                if "partial archived" in tags:
-                    dx.api.project_remove_tags(
-                        project_id,
-                        input_params={
-                            "tags": ["partial archived", "fully archived"],
-                        },
-                    )
-                    add_tag_to_project("fully archived", project_id)
-                elif "fully archived" in tags:  # correctly tagged
-                    continue
-                else:
-                    add_tag_to_project("fully archived", project_id)
+            if "partial archived" in tags:
+                dx.api.project_remove_tags(
+                    project_id,
+                    input_params={
+                        "tags": ["partial archived", "fully archived"],
+                    },
+                )
+                _add_tag_to_project("fully archived", project_id)
+            elif "fully archived" in tags:  # correctly tagged
+                continue
+            else:
+                _add_tag_to_project("fully archived", project_id)
         else:
             # all files are live, no tagging needed
             continue
 
 
-def make_datetime_format(modified_epoch: str) -> dt.datetime:
+def _make_datetime_format(modified_epoch: str) -> dt.datetime:
     """
     Function to turn modified epoch (returned by DNANexus)
     into readable datetime format
@@ -546,6 +315,7 @@ def get_old_tar_and_notify(
     tar_month: int,
     slack: SlackClass,
     project_52: str = "project-FpVG0G84X7kzq58g19vF1YJQ",
+    debug: bool = False,
 ) -> None:
     """
     Function to get tar which are not modified in the last 3 months
@@ -574,20 +344,25 @@ def get_old_tar_and_notify(
                 },
             },
             project=project_52,
+            limit=5 if debug else None,
         )
     )
 
     # list of tar files not modified in the last 3 months
     filtered_result = [
-        x for x in result if older_than(tar_month, x["describe"]["modified"])
+        x for x in result if _older_than(tar_month, x["describe"]["modified"])
     ]
+
+    if not filtered_result:
+        # no .tar older than tar_month
+        return None
 
     id_results = [
         (x["id"], x["describe"]["folder"], x["describe"]["name"])
         for x in filtered_result
     ]
 
-    dates = [make_datetime_format(d["describe"]["modified"]) for d in filtered_result]
+    dates = [_make_datetime_format(d["describe"]["modified"]) for d in filtered_result]
 
     # earliest date among the list of tars
     min_date = min(dates).strftime("%Y-%m-%d")
